+++ packages/perseus-core/src/utils/split-perseus-renderer.test.ts
@@ -384,7 +384,8 @@ describe("splitPerseusRenderer", () => {
         });
     });
 
-    it("upgrades widgets before splitting", () => {
+    // No longer a relevant test, waiting for https://github.com/Khan/perseus/pull/2419  to land before removing
+    it.skip("upgrades widgets before splitting", () => {
         // Arrange
         const v0RadioOptions = {
             choices: [
@@ -424,7 +425,6 @@ describe("splitPerseusRenderer", () => {
                                 content: "Choice 2",
                             },
                         ],
-                        displayCount: null,
                         randomize: false,
                         hasNoneOfTheAbove: false,
                         multipleSelect: false,